Add some unittests. And ... Add failsafe plugin to run some tests during IntegrationTest phase. Rename AppTest to AppIT

+++ src/main/java/com/capgemini/HotelMolvenaGr1/service/BookingService.java
@@ -31,13 +31,17 @@ public class BookingService {
 
     }
 
-    public void deleteBooking(Long id){
-        Booking victum = this.bookingRepository.findOne(id);
-        for (Room r : victum.getRooms()){
-            r.getBookingOwner().remove(victum);
-            victum.getRooms().remove(r);
-        }
-        this.bookingRepository.delete(id);
-    }
+    @Transactional
+    public void deleteBooking(final long id){
 
+        if(this.bookingRepository.exists(id)) {
+
+            Booking victim = this.bookingRepository.findOne(id);
+            for (Room r : victim.getRooms()){
+                r.getBookingOwner().remove(victim);
+                victim.getRooms().remove(r);
+            }
+            this.bookingRepository.delete(id);
+        }
+    }
 }
